refactor: move studentForm submit function to custom hook and fix the loading bug
